Update dashboard page to use global layout with common sidebar and top nav; simplify dashboard content
